storybook - move stories out of core src to reduce dist size
