+++ stories/responsive-chart-wrapper.js
