+++ stories/scatter.stories.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import ScatterChart from '../components/scatter-chart'
-import scatterChartData from '../shared/data/scatter-chart-data'
+import ScatterChart from '../src/components/scatter-chart'
+import scatterChartData from '../src/shared/data/scatter-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
